+++ NEWS.md
@@ -4,3 +4,4 @@
 
 * Fix exponentiation with negative base (#2).
 * Fix formatting when the value equals zero (#3).
+* Fix formatting when the error equals zero (#5).